fix recyclerview fix fragment
can add names to list with non custom view

+++ app/src/main/java/com/example/sprintchallenge3/CreateItem.kt
@@ -1,3 +1,3 @@
 package com.example.sprintchallenge3
 
-class CreateItem (val name: String, val rating: Int)
+class CreateItem (val name: String, val rating: Float)
